requirements.txt
aiogram==3.17.0
ccxt==4.4.16
pandas==2.2.3
python-dotenv==1.0.0
flask==3.1.0
aiohttp==3.11.11